Arthur:finished

+++ app/Tasks/CalculateExplosion/GetMassFromAtomsTask.php
@@ -7,7 +7,7 @@ use App\Constants\Constants;
 class GetMassFromAtomsTask
 {
 
-    public function run(int $atomsAmount, string $coreMaterial): float|int
+    public function run(float $atomsAmount, string $coreMaterial): float|int
     {
         $atomicMass = $coreMaterial == Constants::$CORE_MATERIAL['Uranium'] ? Constants::$URANIUM_ATOMIC_MASS : Constants::$PLUTONIUM_ATOMIC_MASS;
         $mol = $atomsAmount / Constants::$AVOGADRO;
